fix: Move packet from proto converter
